move tendencies to common and make necessary changes
