+++ src/CalibrateCMP/CalibrateCMP.jl
@@ -21,7 +21,6 @@ const FT = Float64
 
 import ..Common as CO
 import ..K1DModel as KD
-import ..K1DColSedModel as KCS
 
 include("ReferenceStats.jl")
 include("ReferenceModels.jl")
