--- src/K1D_col_sed/K1DColSedModel.jl
@@ -1,23 +1,0 @@
-module K1DColSedModel
-
-import OrdinaryDiffEq
-import NCDatasets
-import SpecialFunctions as SF
-import ClimaCore
-import CloudMicrophysics as CM
-
-const CC = ClimaCore
-const ODE = OrdinaryDiffEq
-const CMP = CM.Parameters
-const CM0 = CM.Microphysics0M
-const CM1 = CM.Microphysics1M
-const CM2 = CM.Microphysics2M
-
-import ..Common as CO
-import ..K1DModel as K1D
-
-include("./ode_utils.jl")
-include("./initial_condition.jl")
-include("./tendency.jl")
-
-end
