+++ src/K2DModel/tendency.jl
@@ -1,10 +1,5 @@
 """
-   Different components of the ODE `rhs!` function,
-   depending on the moisture and precipitation types.
-"""
-
-"""
-    Precompute momentum field
+    Prescribed momentum flux as a function of time
 """
 @inline function prescribed_momentum_helper(x, z, t, w1, t1, width, height)
     time_scale = t < t1 ? sin(pi * t / t1) : 0.0
@@ -124,20 +119,20 @@ end
 
     @. dY.ρq_rai +=
         -∂(
-            (aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + aux.precip_velocities.term_vel_rai) *
+            (aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + CC.Geometry.WVector(If(aux.precip_velocities.term_vel_rai) * FT(-1))) *
             If(Y.ρq_rai),
         )
     @. dY.ρq_sno +=
         -∂(
-            (aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + aux.precip_velocities.term_vel_sno) *
+            (aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + CC.Geometry.WVector(If(aux.precip_velocities.term_vel_sno) * FT(-1))) *
             If(Y.ρq_sno),
         )
 
     fcc = CC.Operators.FluxCorrectionC2C(bottom = CC.Operators.Extrapolate(), top = CC.Operators.Extrapolate())
     @. dY.ρq_rai +=
-        fcc((aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + aux.precip_velocities.term_vel_rai), Y.ρq_rai)
+        fcc((aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + CC.Geometry.WVector(If(aux.precip_velocities.term_vel_rai) * FT(-1))), Y.ρq_rai)
     @. dY.ρq_sno +=
-        fcc((aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + aux.precip_velocities.term_vel_sno), Y.ρq_sno)
+        fcc((aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + CC.Geometry.WVector(If(aux.precip_velocities.term_vel_sno) * FT(-1))), Y.ρq_sno)
 
     hdiv = CC.Operators.WeakDivergence()
     @. dY.ρq_rai += -hdiv(aux.prescribed_velocity.ρu / aux.moisture_variables.ρ * Y.ρq_rai)
@@ -155,20 +150,20 @@ end
 
     @. dY.N_rai +=
         -∂(
-            (aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + aux.precip_velocities.term_vel_N_rai) *
+            (aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + CC.Geometry.WVector(If(aux.precip_velocities.term_vel_N_rai) * FT(-1))) *
             If(Y.N_rai),
         )
     @. dY.ρq_rai +=
         -∂(
-            (aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + aux.precip_velocities.term_vel_rai) *
+            (aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + CC.Geometry.WVector(If(aux.precip_velocities.term_vel_rai) * FT(-1))) *
             If(Y.ρq_rai),
         )
 
     fcc = CC.Operators.FluxCorrectionC2C(bottom = CC.Operators.Extrapolate(), top = CC.Operators.Extrapolate())
     @. dY.N_rai +=
-        fcc((aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + aux.precip_velocities.term_vel_N_rai), Y.N_rai)
+        fcc((aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + CC.Geometry.WVector(If(aux.precip_velocities.term_vel_N_rai) * FT(-1))), Y.N_rai)
     @. dY.ρq_rai +=
-        fcc((aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + aux.precip_velocities.term_vel_rai), Y.ρq_rai)
+        fcc((aux.prescribed_velocity.ρw / If(aux.moisture_variables.ρ) + CC.Geometry.WVector(If(aux.precip_velocities.term_vel_rai) * FT(-1))), Y.ρq_rai)
 
 
     hdiv = CC.Operators.WeakDivergence()
