+++ src/Kinematic1D.jl
@@ -4,7 +4,6 @@ module Kinematic1D
 include("Common/Common.jl")
 include("BoxModel/BoxModel.jl")
 include("K1DModel/K1DModel.jl")
-include("K1D_col_sed/K1DColSedModel.jl")
 include("K2DModel/K2DModel.jl")
 include("CalibrateCMP/CalibrateCMP.jl")
