+++ test/create_parameters.jl
@@ -1,3 +1,4 @@
+import Kinematic1D.Common as CO
 import Kinematic1D.K1DModel as KID
 import CLIMAParameters as CP
 import CloudMicrophysics as CM
@@ -6,7 +7,7 @@ import Thermodynamics as TD
 #! format: off
 function override_toml_dict(
     out_dir::String,
-    toml_dict::CP.AbstractTOMLDict,
+    toml_dict::CP.AbstractTOMLDict;
     w1 = 2.0,
     t1 = 600.0,
     p0 = 100700.0,
@@ -107,9 +108,22 @@ function create_thermodynamics_parameters(toml_dict)
     return thermo_params
 end
 
+function create_common_parameters(toml_dict)
+    FTD = CP.float_type(toml_dict)
+    aliases = ["precip_sources", "precip_sinks", "prescribed_Nd"]
+    pairs = CP.get_parameter_values!(toml_dict, aliases, "Common")
+    common_params = CO.Parameters.CommonParameters{FTD}(; pairs...)
+    if !isbits(common_params)
+        print(common_params)
+        @warn "The parameter set SHOULD be isbits in order to be stack-allocated."
+    end
+    return common_params
+end
+Base.broadcastable(x::CO.Parameters.CommonParameters) = Ref(x)
+
 function create_kid_parameters(toml_dict)
     FTD = CP.float_type(toml_dict)
-    aliases = ["w1", "t1", "p0", "precip_sources", "precip_sinks", "qtot_flux_correction", "prescribed_Nd", "r_dry", "std_dry", "κ"]
+    aliases = ["w1", "t1", "p0", "qtot_flux_correction", "r_dry", "std_dry", "κ"]
     pairs = CP.get_parameter_values!(toml_dict, aliases, "Kinematic1D")
     kid_params = KID.Parameters.Kinematic1DParameters{FTD}(; pairs...)
     if !isbits(kid_params)
