+++ test/experiments/KiD_col_sed_driver/run_KiD_col_sed_simulation.jl
@@ -3,11 +3,12 @@ import ClimaCore as CC
 import CLIMAParameters as CP
 import CloudMicrophysics.Parameters as CMP
 import Kinematic1D
-import Kinematic1D.K1DModel as KID
-import Kinematic1D.K1DColSedModel as KCS
+import Kinematic1D.Common as CO
+import Kinematic1D.K1DModel as K1D
 
 include(joinpath(pkgdir(Kinematic1D), "test", "create_parameters.jl"))
-include("plotting_utils.jl")
+include(joinpath(pkgdir(Kinematic1D), "test", "plotting_utils.jl"))
+# include("plotting_utils.jl")
 
 function run_KiD_col_sed_simulation(::Type{FT}, opts) where {FT}
 
@@ -32,15 +33,20 @@ function run_KiD_col_sed_simulation(::Type{FT}, opts) where {FT}
     toml_dict = override_toml_dict(
         path,
         default_toml_dict,
-        Int(opts["precip_sources"]),
-        Int(opts["precip_sinks"]),
-        FT(opts["prescribed_Nd"]),
+        precip_sources = Int(opts["precip_sources"]),
+        precip_sinks = Int(opts["precip_sinks"]),
+        prescribed_Nd = FT(opts["prescribed_Nd"]),
     )
     toml_dict["νc_SB2006"]["value"] = opts["k"]
-    # Create Kinematic1D model parameters
+    # Create Thermodynamics.jl and Kinematic1D model parameters
     # (some of the CloudMicrophysics.jl parameters structs are created later based on model choices)
+    common_params = create_common_parameters(toml_dict)
     kid_params = create_kid_parameters(toml_dict)
+    thermo_params = create_thermodynamics_parameters(toml_dict)
+    air_params = CMP.AirProperties(FT, toml_dict)
+    activation_params = CMP.AerosolActivationParameters(FT, toml_dict)
 
+    moisture = CO.get_moisture_type(FT, "NonEquilibriumMoisture", toml_dict)
     precip = CO.get_precipitation_type(
         FT, 
         precipitation_choice, 
@@ -49,22 +55,23 @@ function run_KiD_col_sed_simulation(::Type{FT}, opts) where {FT}
         sedimentation_choice = sedimentation_choice)
 
     # Initialize the timestepping struct
-    TS = KID.TimeStepping(FT(opts["dt"]), FT(opts["dt_output"]), FT(opts["t_end"]))
+    TS = CO.TimeStepping(FT(opts["dt"]), FT(opts["dt_output"]), FT(opts["t_end"]))
 
     # Create the coordinates
-    space, face_space = KID.make_function_space(FT, FT(opts["z_min"]), FT(opts["z_max"]), opts["n_elem"])
+    space, face_space = K1D.make_function_space(FT, FT(opts["z_min"]), FT(opts["z_max"]), opts["n_elem"])
     coord = CC.Fields.coordinate_field(space)
     face_coord = CC.Fields.coordinate_field(face_space)
 
     # Initialize the netcdf output Stats struct
     fname = joinpath(path, "Output.nc")
-    Stats = KID.NetCDFIO_Stats(
+    Stats = K1D.NetCDFIO_Stats(
         fname,
         1.0,
         parent(face_coord),
         parent(coord),
         output_profiles = Dict(
             :ρ => "density",
+            :q_tot => "q_tot",
             :q_liq => "q_liq",
             :q_rai => "q_rai",
             :N_liq => "N_liq",
@@ -74,24 +81,36 @@ function run_KiD_col_sed_simulation(::Type{FT}, opts) where {FT}
 
     # Create the initial condition profiles
     init =
-        map(coord -> KCS.init_1d_column(FT, opts["qt"], opts["prescribed_Nd"], opts["k"], opts["rhod"], coord.z), coord)
+        map(coord -> K1D.init_1d_column(FT, thermo_params, opts["qt"], opts["prescribed_Nd"], opts["k"], opts["rhod"], coord.z), coord)
 
     # Create state vector and apply initial condition
-    Y = KCS.initialise_state(precip, init)
+    Y = CO.initialise_state(moisture, precip, init)
 
     # Create aux vector and apply initial condition
-    aux = KCS.initialise_aux(FT, init, kid_params, TS, Stats, face_space)
+    aux = K1D.initialise_aux(
+        FT,
+        init,
+        common_params,
+        kid_params,
+        thermo_params,
+        air_params,
+        activation_params,
+        TS,
+        Stats,
+        face_space,
+        moisture,
+    )
 
     # Output the initial condition
-    KID.KiD_output(aux, 0.0)
+    K1D.KiD_output(aux, 0.0)
 
     # Define callbacks for output
-    callback_io = ODE.DiscreteCallback(KID.condition_io, KID.affect_io!; save_positions = (false, false))
+    callback_io = ODE.DiscreteCallback(K1D.condition_io, K1D.affect_io!; save_positions = (false, false))
     callbacks = ODE.CallbackSet(callback_io)
 
     # Collect all the tendencies into rhs function for ODE solver
     # based on model choices for the solved equations
-    ode_rhs! = KCS.make_rhs_function(precip)
+    ode_rhs! = K1D.make_rhs_function_col_sed(moisture, precip)
 
     # Solve the ODE operator
     problem = ODE.ODEProblem(ode_rhs!, Y, (FT(opts["t_ini"]), FT(opts["t_end"])), aux)
@@ -106,9 +125,8 @@ function run_KiD_col_sed_simulation(::Type{FT}, opts) where {FT}
     )
 
     # Some basic plots
-    plot_folder =
-        joinpath(pkgdir(Kinematic1D), string("test/experiments/KiD_col_sed_driver/", output_folder, "/figures/"))
-    plot_timeheight(string(output_folder, "/Output.nc"), output = plot_folder)
+    plot_folder = string("experiments/KiD_col_sed_driver/", output_folder, "/figures/")
+    plot_timeheight(string("experiments/KiD_col_sed_driver/", output_folder, "/Output.nc"), output = plot_folder)
 
     return solver
 end
