+++ test/experiments/KiD_driver/run_KiD_simulation.jl
@@ -3,7 +3,8 @@ import ClimaCore as CC
 import CLIMAParameters as CP
 import CloudMicrophysics.Parameters as CMP
 import Kinematic1D
-import Kinematic1D.K1DModel as KID
+import Kinematic1D.Common as CO
+import Kinematic1D.K1DModel as K1D
 
 include(joinpath(pkgdir(Kinematic1D), "test", "create_parameters.jl"))
 include(joinpath(pkgdir(Kinematic1D), "test", "plotting_utils.jl"))
@@ -36,21 +37,22 @@ function run_KiD_simulation(::Type{FT}, opts) where {FT}
     toml_dict = override_toml_dict(
         path,
         default_toml_dict,
-        FT(opts["w1"]),
-        FT(opts["t1"]),
-        FT(opts["p0"]),
-        Int(opts["precip_sources"]),
-        Int(opts["precip_sinks"]),
-        Int(opts["qtot_flux_correction"]),
-        FT(opts["prescribed_Nd"]),
-        FT(opts["r_dry"]),
-        FT(opts["std_dry"]),
-        FT(opts["kappa"]),
+        w1 = FT(opts["w1"]),
+        t1 = FT(opts["t1"]),
+        p0 = FT(opts["p0"]),
+        precip_sources = Int(opts["precip_sources"]),
+        precip_sinks = Int(opts["precip_sinks"]),
+        qtot_flux_correction = Int(opts["qtot_flux_correction"]),
+        prescribed_Nd = FT(opts["prescribed_Nd"]),
+        r_dry = FT(opts["r_dry"]),
+        std_dry = FT(opts["std_dry"]),
+        kappa = FT(opts["kappa"]),
     )
     # Create Thermodynamics.jl and Kinematic1D model parameters
     # (some of the CloudMicrophysics.jl parameters structs are created later based on model choices)
-    thermo_params = create_thermodynamics_parameters(toml_dict)
+    common_params = create_common_parameters(toml_dict)
     kid_params = create_kid_parameters(toml_dict)
+    thermo_params = create_thermodynamics_parameters(toml_dict)
     air_params = CMP.AirProperties(FT, toml_dict)
     activation_params = CMP.AerosolActivationParameters(FT, toml_dict)
 
@@ -64,29 +66,30 @@ function run_KiD_simulation(::Type{FT}, opts) where {FT}
     )
 
     # Initialize the timestepping struct
-    TS = KID.TimeStepping(FT(opts["dt"]), FT(opts["dt_output"]), FT(opts["t_end"]))
+    TS = CO.TimeStepping(FT(opts["dt"]), FT(opts["dt_output"]), FT(opts["t_end"]))
 
     # Create the coordinates
-    space, face_space = KID.make_function_space(FT, FT(opts["z_min"]), FT(opts["z_max"]), opts["n_elem"])
+    space, face_space = K1D.make_function_space(FT, FT(opts["z_min"]), FT(opts["z_max"]), opts["n_elem"])
     coord = CC.Fields.coordinate_field(space)
     face_coord = CC.Fields.coordinate_field(face_space)
 
     # Initialize the netcdf output Stats struct
     fname = joinpath(path, "Output.nc")
-    Stats = KID.NetCDFIO_Stats(fname, 1.0, parent(face_coord), parent(coord))
+    Stats = K1D.NetCDFIO_Stats(fname, 1.0, parent(face_coord), parent(coord))
 
     # Solve the initial value problem for density profile
-    ρ_profile = KID.ρ_ivp(FT, kid_params, thermo_params)
+    ρ_profile = K1D.ρ_ivp(FT, kid_params, thermo_params)
     # Create the initial condition profiles
-    init = map(coord -> KID.init_1d_column(FT, kid_params, thermo_params, ρ_profile, coord.z), coord)
+    init = map(coord -> K1D.init_1d_column(FT, common_params, kid_params, thermo_params, ρ_profile, coord.z), coord)
 
     # Create state vector and apply initial condition
-    Y = KID.initialise_state(moisture, precip, init)
+    Y = CO.initialise_state(moisture, precip, init)
 
     # Create aux vector and apply initial condition
-    aux = KID.initialise_aux(
+    aux = K1D.initialise_aux(
         FT,
         init,
+        common_params,
         kid_params,
         thermo_params,
         air_params,
@@ -98,15 +101,15 @@ function run_KiD_simulation(::Type{FT}, opts) where {FT}
     )
 
     # Output the initial condition
-    KID.KiD_output(aux, 0.0)
+    K1D.KiD_output(aux, 0.0)
 
     # Define callbacks for output
-    callback_io = ODE.DiscreteCallback(KID.condition_io, KID.affect_io!; save_positions = (false, false))
+    callback_io = ODE.DiscreteCallback(K1D.condition_io, K1D.affect_io!; save_positions = (false, false))
     callbacks = ODE.CallbackSet(callback_io)
 
     # Collect all the tendencies into rhs function for ODE solver
     # based on model choices for the solved equations
-    ode_rhs! = KID.make_rhs_function(moisture, precip)
+    ode_rhs! = K1D.make_rhs_function(moisture, precip)
 
     # Solve the ODE operator
     problem = ODE.ODEProblem(ode_rhs!, Y, (FT(opts["t_ini"]), FT(opts["t_end"])), aux)
@@ -127,7 +130,7 @@ function run_KiD_simulation(::Type{FT}, opts) where {FT}
 
         z_centers = parent(CC.Fields.coordinate_field(space))
         plot_final_aux_profiles(z_centers, aux, precip, output = plot_folder)
-        plot_animation(z_centers, solver, aux, moisture, precip, KID, output = plot_folder)
+        plot_animation(z_centers, solver, aux, moisture, precip, K1D, output = plot_folder)
         plot_timeheight(string("experiments/KiD_driver/", output_folder, "/Output.nc"), output = plot_folder)
     end
